fix: Updated data types of foreign keys

+++ .gen/ezsplit_dev/public/model/items.go
@@ -13,7 +13,7 @@ import (
 
 type Items struct {
 	ID        int32 `sql:"primary_key"`
-	ReceiptID int32
+	ReceiptID int64
 	Price     *int32
 	Name      *string
 	CreatedAt time.Time
